backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
opencv-python-headless
pytesseract
python-multipart
python-dateutil
pydantic
pandas